requirements.txt
appdirs>=1.4.4
black>=23.1.0
click>=8.1.3
dnspython>=2.3.0
email-validator>=1.3.1
fastapi>=0.95.0
flake8>=6.0.0
h11>=0.14.0
idna>=3.4
mccabe>=0.7.0
motor>=3.1.1
mypy-extensions>=1.0.0
pathspec>=0.11.1
pycodestyle>=2.10.0
pydantic<=2.0
pyflakes>=3.0.1
pymongo>=4.3.3
regex>=2023.3.23
starlette>=0.26.1
toml>=0.10.2
typing_extensions>=4.5.0
uvicorn>=0.21.1
python-multipart>=0.0.6
aiofiles>=23.1.0
PyJWT>=2.6.0
fastapi-another-jwt-auth>=0.1.7
redis>=4.5.3
bcrypt>=4.0.1
python-dotenv>=1.0.0

# dashboard
a2wsgi>=1.10.10
blinker>=1.9.0
certifi>=2025.10.5
charset-normalizer>=3.4.3
click>=8.3.0
dash>=3.2.0
dash-bootstrap-components>=2.0.4
flask>=3.1.2
idna>=3.10
importlib-metadata>=8.7.0
itsdangerous>=2.2.0
jinja2>=3.1.6
markupsafe>=3.0.3
narwhals>=2.7.0
nest-asyncio>=1.6.0
numpy>=2.3.3
packaging>=25.0
pandas>=2.3.3
plotly>=6.3.1
psycopg2-binary>=2.9.10
python-dateutil>=2.9.0.post
pytz>=2025.2
requests>=2.32.5
retrying>=1.4.2
schedule>=1.2.2
typing-extensions>=4.15.0
tzdata>=2025.2
urllib3>=2.5.0
werkzeug>=3.1.3

# for iridium-toolkit
crcmod
